refactor+fix: cache leaderboard routes
we keep track of the lowest value for the most_kills and most_damage_dealt leaderboards. If the max value in the game is greater than that, we recalculate the leaderboard; otherwise we serve the cached value, which is cached for one day.
For the other leaderboards, I didn't find a way to intelligently invalidate them, so they are just cached for 6 minutes
I didn't want to spend a lot of time prematurely optimizing this without having any actual bottlenecks..