fix: first grid query always failing
So, all objects grid query ID are initialized to 0, and so is the first grid query...

This causes the first grid query of every game to return an empty array

and guess what that query is on cobalt?

The top twins bunker door checking for stairs to figure out which layer it should be in.
That query failing causes the door to stay on layer 1 (underground) instead of going to layer 3 (stair -> underground).
Wich will make the client ceiling reveal raycasting not check for it!