Update stats.ts
The schema would return NaN for mapid if it was not a number, which resulted in an error when trying to retrieve values from the database (mapid had to be int). All this did was return an error 500 instead of 400, but this should fix it. Also first time doing a github fork/push thing so sorry if i did this wrong